Refactor(helm): extract Provider, constructors, and dry-run   ctx into provider.go
Move the Provider singleton, NewProvider/NewProviderWithConfig
   constructors, DefaultCacheTTLConfig, the dry-run context helpers
  (WithDryRun/isDryRun), and the small log/cache key utilities out of
  helm.go into a dedicated provider.go.

No behavior change. All 240 unit specs remain green.
